Group reducers to global state file

+++ src/app/modules/users/components/view/view.component.ts
@@ -1,46 +1,49 @@
-import {AfterViewInit, Component, OnInit, ViewChild} from '@angular/core';
+import {AfterViewInit, Component, OnDestroy, OnInit, ViewChild} from '@angular/core';
 import {User} from "../../../../models/user";
 import {UserService} from "../../../../services/user/user.service";
 import {Router} from "@angular/router";
 import {MatPaginator} from "@angular/material/paginator";
 import {MatTableDataSource} from "@angular/material/table";
 import {select, Store} from "@ngrx/store";
-import {usersList} from "../../../../states/users/users.selectors";
+import {selectUsersList} from "../../../../states/users/users.selectors";
 import {UsersActions} from "../../../../states/users/users.actions";
+import {Subject, takeUntil} from "rxjs";
+import {AppState} from "../../../../states/app.states";
 
 @Component({
   selector: 'app-view',
   templateUrl: './view.component.html',
   styleUrls: ['./view.component.css'],
 })
-export class ViewComponent implements OnInit, AfterViewInit {
+export class ViewComponent implements OnInit, AfterViewInit, OnDestroy {
 
-  private userService: UserService;
-  private router: Router;
+  constructor(
+    private userService: UserService,
+    private router: Router,
+    private store: Store<AppState>
+  ) {}
+  private destroy$ = new Subject<void>();
 
-  constructor(userService: UserService, router: Router, private store: Store) {
-    this.store = store;
-    this.router = router;
-    this.userService = userService;
-  }
-
-
-  // users$ = this.store.select(usersList);
   displayedColumns: string[] = ['id', 'first_name', 'last_name', 'action'];
-  dataSource = new MatTableDataSource<User>([]);
+  dataSource = new MatTableDataSource<User>();
 
   @ViewChild(MatPaginator) paginator!: MatPaginator;
 
   ngOnInit(): void {
-    this.store.pipe(select(usersList)).subscribe((users) => {
+    this.store.pipe(select(selectUsersList), takeUntil(this.destroy$)).subscribe((users) => {
       this.dataSource.data = users;
     });
+
     this.store.dispatch(UsersActions.loadUsers());
-    // this.dataSource.data = this.userService.getAll().reverse();
   }
 
   ngAfterViewInit() {
     this.dataSource.paginator = this.paginator;
+  }
+
+  ngOnDestroy(): void {
+    this.destroy$.next();
+    this.destroy$.complete();
   }
 
   create() {
@@ -52,6 +55,8 @@ export class ViewComponent implements OnInit, AfterViewInit {
   }
 
   delete(id: number) {
-    this.userService.delete(id);
+    if (confirm('Do you really want to delete this user? Warning: This is irreversible.')) {
+      this.userService.delete(id);
+    }
   }
 }
